+++ NEWS.md
@@ -1,3 +1,10 @@
-# pannotator (version 1.0.0.3)
+
+# pannotator (version 1.0.1)
+
+* Replaced leaflet.extras dependency.
+* Reordered new dropdowns to appear at top.
+* Changed load kmz button to use different progress bar saving screen space.
+
+# pannotator (version 1.0.0.4)
 
 * Initial CRAN submission.